+++ source/renderer/app/components/wallet/wallet-import/WalletSelectImportDialog.js
@@ -293,9 +293,13 @@ export default class WalletSelectImportDialog extends Component<Props> {
             <div className={styles.walletsContainer}>
               {walletsWithNames.map(wallet => {
                 const isDuplicate = previousWalletId === wallet.id;
+                const rowClasses = classNames([
+                  styles.walletsRow,
+                  'namedWalletsRow',
+                ]);
                 const walletRow = (
                   <div
-                    className={styles.walletsRow}
+                    className={rowClasses}
                     key={`${wallet.id}-${wallet.index}`}
                   >
                     <div className={styles.walletsCounter}>
@@ -336,9 +340,13 @@ export default class WalletSelectImportDialog extends Component<Props> {
 
               {walletsWithoutNames.map(wallet => {
                 const isDuplicate = previousWalletId === wallet.id;
+                const rowClasses = classNames([
+                  styles.walletsRow,
+                  'unnamedWalletsRow',
+                ]);
                 const walletRow = (
                   <div
-                    className={styles.walletsRow}
+                    className={rowClasses}
                     key={`${wallet.id}-${wallet.index}`}
                   >
                     <div className={styles.walletsCounter}>