+++ source/renderer/app/containers/wallet/dialogs/WalletImportDialogContainer.js
@@ -1,19 +1,33 @@
 // @flow
 import React, { Component } from 'react';
 import { observer, inject } from 'mobx-react';
+import { map } from 'lodash';
 import WalletImportFileDialog from '../../../components/wallet/wallet-import/WalletImportFileDialog';
 import WalletSelectImportDialog from '../../../components/wallet/wallet-import/WalletSelectImportDialog';
 import { isValidWalletName } from '../../../utils/validations';
 import { IMPORT_WALLET_STEPS } from '../../../config/walletRestoreConfig';
 import type { InjectedProps } from '../../../types/injectedPropsType';
 import type { ImportFromOption } from '../../../types/walletExportTypes';
+import { MAX_ADA_WALLETS_COUNT } from '../../../config/numbersConfig';
+import { WalletImportStatuses } from '../../../types/walletExportTypes';
 
 type Props = InjectedProps;
 
+type State = {
+  existingWalletsCount: number,
+};
+
 @inject('stores', 'actions')
 @observer
-export default class WalletImportDialogContainer extends Component<Props> {
+export default class WalletImportDialogContainer extends Component<
+  Props,
+  State
+> {
   static defaultProps = { actions: null, stores: null };
+
+  state = {
+    existingWalletsCount: this.props.stores.wallets.all.length,
+  };
 
   get containers() {
     return {
@@ -72,27 +86,33 @@ export default class WalletImportDialogContainer extends Component<Props> {
     } = walletMigration;
 
     const { openExternalLink } = app;
+    let walletsCount =
+      this.state.existingWalletsCount + pendingImportWalletsCount;
+    map(exportedWallets, wallet => {
+      if (
+        wallet.import.status === WalletImportStatuses.COMPLETED ||
+        wallet.import.status === WalletImportStatuses.RUNNING
+      ) {
+        walletsCount++;
+      }
+    });
+    const isMaxNumberOfWalletsReached = walletsCount >= MAX_ADA_WALLETS_COUNT;
 
-    const {
-      importWalletStep,
-    } = stores.wallets;
+    const { importWalletStep } = stores.wallets;
 
-    const {
-      restoreWalletClose,
-      restoreWalletChangeStep,
-    } = actions.wallets;
+    const { restoreWalletClose, restoreWalletChangeStep } = actions.wallets;
 
     const stepId = IMPORT_WALLET_STEPS[importWalletStep];
 
     const CurrentContainer = this.containers[stepId];
 
     return (
-        <CurrentContainer
-          onContinue={() => restoreWalletChangeStep.trigger()}
-          onBack={() => restoreWalletChangeStep.trigger(true)}
-          onClose={() => restoreWalletClose.trigger()}
-        />
-        /* {walletMigrationStep === 1 && (
+      <CurrentContainer
+        onContinue={() => restoreWalletChangeStep.trigger()}
+        onBack={() => restoreWalletChangeStep.trigger(true)}
+        onClose={() => restoreWalletClose.trigger()}
+      />
+      /* {walletMigrationStep === 1 && (
           <WalletImportFileDialog
             isSubmitting={isExportRunning}
             exportSourcePath={exportSourcePath}
@@ -118,6 +138,7 @@ export default class WalletImportDialogContainer extends Component<Props> {
             onWalletNameChange={this.onWalletNameChange}
             onToggleWalletImportSelection={this.onToggleWalletImportSelection}
             onClose={this.onCancel}
+            isMaxNumberOfWalletsReached={isMaxNumberOfWalletsReached}
           />
         )} */
     );