+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 Changelog
 =========
 
+## vNext
+
+### Fixes
+
+- Disabled button on forms when there is nothing to submit ([PR 1998](https://github.com/input-output-hk/daedalus/pull/1998), [PR 2010](https://github.com/input-output-hk/daedalus/pull/2010))
+- Fixed system locale detection ([PR 2009](https://github.com/input-output-hk/daedalus/pull/2009))
+
+### Chores
+
+- Prevented import for more than maximum number of wallets ([PR 2012](https://github.com/input-output-hk/daedalus/pull/2012))
+- Implemented automated tests for wallet migration and import features ([PR 2011](https://github.com/input-output-hk/daedalus/pull/2011))
+
 ## 1.1.0-FC1
 
 ### Features